+++ commons/factory.js
@@ -69,6 +69,14 @@ exports.getUserToken = async function (user, tenant) {
   return authentication.encode(user, tenant);
 };
 
+exports.getDeviceToken = async function (device, tenant) {
+  const Tenant = mongoose.dbs["catalog"].model("Tenant");
+  if (!tenant)
+    if (!tenant) tenant = await Tenant.findById(process.env.DEFAULT_TENANT);
+  return authentication.encodeDevice(device, tenant);
+};
+
+
 exports.createTenant = async function (id, organization, address, email, phone, admin_username, admin_password) {
   const Tenant = mongoose.dbs["catalog"].model("Tenant");
   let tenant = await Tenant.findOne({ _id: id });
@@ -90,21 +98,21 @@ exports.createTenant = async function (id, organization, address, email, phone, 
   return await Tenant.findById(tenant._id);
 };
 
-exports.createUser = async function (username, password, type, fieldmask, email, tenant,validityPasswordDays,createdPassword) {
+exports.createUser = async function (username, password, type, fieldmask, email, tenant, validityPasswordDays, createdPassword) {
   const Tenant = mongoose.dbs["catalog"].model("Tenant");
   if (!tenant) tenant = await Tenant.findById(process.env.DEFAULT_TENANT);
-  if (tenant.passwordhash == true || tenant.passwordhash == "true") {
+  if (tenant.passwordhash != false && tenant.passwordhash != "false") {
     password = bcrypt.hashSync(password, 8);
   }
   const User = mongoose.dbs[tenant.database].model("User");
-  const req = { 
+  const req = {
     username: username,
     password: password,
     fieldmask: fieldmask,
-    email: email||username+"@gmail.com",
+    email: email || username + "@gmail.com",
     type: type || UserRoles.provider,
-    validityPasswordDays:validityPasswordDays||0,
-    createdPassword:createdPassword
+    validityPasswordDays: validityPasswordDays || 0,
+    createdPassword: createdPassword
   };
   let user = new User(req);
   await user.save();
@@ -130,13 +138,13 @@ exports.createRole = async function (name, defaultAction, actions, description, 
   return role._doc;
 };
 
-exports.createCrud = async function (create,read,update,deleteAction) {
+exports.createCrud = async function (create, read, update, deleteAction) {
   const body = {}
-  if(create===true||create===false)body.create= create;
-  if(read)body.read= read;
-  if(update)body.update= update;
-  if(deleteAction)body.delete= deleteAction;
-  
+  if (create === true || create === false) body.create = create;
+  if (read) body.read = read;
+  if (update) body.update = update;
+  if (deleteAction) body.delete = deleteAction;
+
   return body;
 };
 
@@ -182,7 +190,7 @@ exports.createDefaultRoles = async function (tenant) {
   }
   ];
   let roleDocs = []
-  for(let i=0;i<body.length;i++){
+  for (let i = 0; i < body.length; i++) {
     let role = new Role(body[i]);
     await role.save();
     roleDocs.push(role._doc)
@@ -262,7 +270,7 @@ exports.createMetadata = async function (name, description, type, range) {
     name: name || "metadata-name-" + this.uuid(),
     description: description || "description metadata " + this.uuid(),
     type: type || MetadataTypes.scalar,
-    range:range||[]
+    range: range || []
   }
   return metadata;
 }
@@ -272,7 +280,7 @@ exports.createField = async function (name, description, type, range) {
     name: name || "field-name-" + this.uuid(),
     description: description || "description field " + this.uuid(),
     type: type || TopicFieldTypes.scalar,
-    range:range||[]
+    range: range || []
   }
   return field;
 }
@@ -339,7 +347,7 @@ exports.createExperiment = async function (name, description, owner, state, star
   const Experiment = mongoose.dbs[tenant.database].model("Experiment");
   const req = {
     _id: name,
-    description: description,    
+    description: description,
     //state: state || ExperimentStateTypes.ongoing,
     state: state || 0,
     startDate: startDate || Date.now(),
@@ -358,9 +366,10 @@ exports.createExperiment = async function (name, description, owner, state, star
   return experiment._doc;
 };
 
-exports.createDevice = async function (name, owner, features, tags, scripts, visibility, tenant) {
+exports.createDevice = async function (name, owner, features, tags, scripts, visibility, tenant, things, token) {
   const Tenant = mongoose.dbs["catalog"].model("Tenant");
   if (!tenant) tenant = await Tenant.findById(process.env.DEFAULT_TENANT);
+  
   const Device = mongoose.dbs[tenant.database].model("Device");
   const req = {
     _id: name,
@@ -368,8 +377,14 @@ exports.createDevice = async function (name, owner, features, tags, scripts, vis
     tags: tags,
     scripts: scripts,
     features: features || [await this.createFeature(name + "-feature", owner)],
-    visibility: visibility,
+    visibility: visibility || VisibilityTypes.private,    
+    things: things
   };
+  if(!token) token = authentication.encodeDevice(req, tenant);
+  if (tenant.passwordhash != false && tenant.passwordhash != "false") {
+    token = bcrypt.hashSync(token, 8);
+  }
+  req.token=token;
   const device = new Device(req);
   await device.save();
   return device._doc;
@@ -579,19 +594,20 @@ exports.createFieldmask = async function (
 exports.createMeasurement = async function (owner, feature, device, thing, tags, samples, startdate, enddate, location, visibility, experiment, tenant) {
   const Tenant = mongoose.dbs["catalog"].model("Tenant");
   if (!tenant) tenant = await Tenant.findById(process.env.DEFAULT_TENANT);
-  if(samples == undefined) samples = [{ values: [10.4], delta: 200 }];
+  if (samples == undefined) samples = [{ values: [10.4], delta: 200 }];
   const Measurement = mongoose.dbs[tenant.database].model("Measurement");
-  const req = { owner: owner,
-                startDate: startdate || Date.now(),
-                endDate: enddate || Date.now(),
-                location: location || { type: "Point", coordinates: [12.123456, 13.1345678] },
-                thing: thing,
-                feature: feature,
-                device: device,
-                samples: samples,
-                tags: tags,
-                visibility: visibility,
-                experiment: experiment
+  const req = {
+    owner: owner,
+    startDate: startdate || Date.now(),
+    endDate: enddate || Date.now(),
+    location: location || { type: "Point", coordinates: [12.123456, 13.1345678] },
+    thing: thing,
+    feature: feature,
+    device: device,
+    samples: samples,
+    tags: tags,
+    visibility: visibility,
+    experiment: experiment
   };
   const id = sha(JSON.stringify(req));
   req._id = id;
@@ -603,17 +619,18 @@ exports.createMeasurement = async function (owner, feature, device, thing, tags,
 exports.createTimesample = async function (owner, values, timestamp, measurement, tenant) {
   const Tenant = mongoose.dbs["catalog"].model("Tenant");
   if (!tenant) tenant = await Tenant.findById(process.env.DEFAULT_TENANT);
-  if(!measurement) {
-      const Measurement = mongoose.dbs[tenant.database].model("Measurement");
-      const feature = await this.createFeature(this.uuid(), owner);
-      const device = await this.createDevice(this.uuid(), owner, [feature]);
-      const thing = await this.createThing(this.uuid(), owner);
-      measurement = await factory.createMeasurement(owner, feature, device, thing, [], []);
+  if (!measurement) {
+    const Measurement = mongoose.dbs[tenant.database].model("Measurement");
+    const feature = await this.createFeature(this.uuid(), owner);
+    const device = await this.createDevice(this.uuid(), owner, [feature]);
+    const thing = await this.createThing(this.uuid(), owner);
+    measurement = await factory.createMeasurement(owner, feature, device, thing, [], []);
   }
   const Timesample = mongoose.dbs[tenant.database].model("Timesample");
-  const req = { measurement: measurement,
-                values: values || [1], 
-                timestamp: timestamp || Date.now()
+  const req = {
+    measurement: measurement,
+    values: values || [1],
+    timestamp: timestamp || Date.now()
   }
   const timesample = new Timesample(req);
   await timesample.save();
@@ -985,7 +1002,7 @@ exports.createDemoContent = async function (tenant) {
       null,
       null,
       null,
-      VisibilityTypes.public,      
+      VisibilityTypes.public,
       null,
       tenant
     )
@@ -1192,4 +1209,13 @@ exports.areEqual = function areEqual(obj1, obj2) {
     return true;
   }
   return false;
+};
+
+exports.HashDeviceToken = async function (tenant, token) {
+  const Tenant = mongoose.dbs["catalog"].model("Tenant");
+  if (!tenant) tenant = await Tenant.findById(process.env.DEFAULT_TENANT);  
+  if (tenant.passwordhash != false && tenant.passwordhash != "false") {
+    token = bcrypt.hashSync(token, 8);
+  }
+  return token;
 };