+++ models/deviceSchema.js
@@ -6,8 +6,10 @@ const MeasurementBufferPolicyTypes = require('../types/measurementBufferPolicyTy
  
 const deviceSchema = new mongoose.Schema({ 
     _id: { type: String, required: "Please, supply an _id" },
+    token: { type: String, unique: true, select: false },
     owner: { type: mongoose.Schema.Types.ObjectId, ref:'User', required: true },
     features: { type: [String], required: "Please, supply a feature array", ref:'Feature' },
+    things: { type: [String], ref:'thing' },
     tags: { type: [String], ref:'Tag' },
     scripts: { type: [String], ref:'Script' },
     visibility: {type: String, enum: VisibilityTypes, default: VisibilityTypes.private },