+++ release-notes.md
@@ -43,3 +43,16 @@
 + Added field "select" in the query of get resources routes to receive only chosen fields
 + Experiment/id/group route added to receive the data based on groups of topics chosen in protocol
 + Removed variables.env to create it from variablesTemplate.env
+
+## Version 0.6.5
+
+2023.02.06
+
++ Now device can post measurements with a token given at device creation without do a login with user credential
+
+## Version 0.6.6
+
+2023.02.13
+
++ Added description and tags fields for items in feature
+